// admin/src/domains/fulfillment/board/graphql/operation-types.ts
export interface ApiFulfillmentStage {
  id: string;
  version: number;
  title: string;
  handle: string;
  sortIndex: number;
  ticketConnection: ApiFulfillmentTicketConnection;
  createdAt: string;
  updatedAt: string;
}

export interface ApiFulfillmentStageEdge {
  cursor: string;
  node: ApiFulfillmentStage;
}

export interface ApiFulfillmentStageConnection {
  edges: ApiFulfillmentStageEdge[];
  pageInfo: ApiPageInfo;
  totalCount: number;
}